server_buddy/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.16
langchain-core==0.1.42
langchain-google-genai==0.0.6
langchain-chroma==0.1.0
chromadb==0.4.24
langchain-community==0.0.32
python-dotenv==1.0.0
pydantic==2.9.2
